Fix compiler warnings.
git-svn-id: https://simulationcraft.googlecode.com/svn/branches/mop@11415 3b4652a1-8050-0410-ac47-3d40261b0f8b

+++ engine/sc_sim.cpp
@@ -1878,7 +1878,7 @@ double sim_t::gauss( double mean,
 timespan_t sim_t::gauss( timespan_t mean,
                          timespan_t stddev )
 {
-  return timespan_t::from_native( gauss( timespan_t::to_native( mean ), timespan_t::to_native( stddev ) ) );
+  return timespan_t::from_native( gauss( (double) timespan_t::to_native( mean ), (double) timespan_t::to_native( stddev ) ) );
 }
 
 // sim_t::get_rng ===========================================================
